backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "fastapi[standard]>=0.115.6",
    "orjson>=3.10.14",
    "pwdlib[argon2]>=0.2.1",
    "pycrypto>=2.6.1",
    "pycryptodome>=3.21.0",
    "pydantic-extra-types>=2.10.2",
    "pydantic-settings>=2.7.1",
    "pyjwt>=2.10.1",
    "python-decouple>=3.8",
    "sqlalchemy>=2.0.37",
    "tzdata>=2024.2",
    "uvicorn[standard]>=0.34.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.1",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.2",
    "taskipy>=1.14.1",
]

[tool.ruff]
line-length = 88
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']
ignore = ["E203", "E501"]

[tool.ruff.format]
preview = true
quote-style = 'single'
docstring-code-format = true

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev app/main.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=app -vvv'
post_test = 'coverage html'
